requirements.txt


Django==3.1.7


dj-database-url==0.5.0

# dj-conf-url
django-admin==2.0.1

gunicorn==20.0.4


# mysql driver
pymysql==1.0.2
# MySQL-python
mysql-connector-python==8.0.23
msql==1.2.2
# mysql==0.0.2
# mysqlclient

# MYSQL Json Field provider
# django-mysql

#whitenoise
whitenoise==5.2.0

cryptography


djangorestframework==3.11.0
django-extensions==3.0.7
django-cors-headers===3.5.0
pyjwt==1.7.1

django-rest-swagger==2.2.0
